Moved most of the logic from the native-common/GestureView control into a straight common/GestureView abstract base component, and changed much of the logic away from using raw events as inputs and into the processed RX GestureState types.  Pulled a bunch of the duplicate logic out of web/GestureView and used the new base class.  Added touchstart/move/end support to web/GestureView to now reuse the touch code out of the common GestureView to get web touch support in GestureView!
