+++ src/native-common/utils/EventHelpers.ts
@@ -6,6 +6,7 @@
  */
 import { Platform } from 'react-native';
 
+import * as EventHelpersCommon from '../../common/utils/EventHelpers';
 import { Types } from '../../common/Interfaces';
 
 import { clone } from './lodashMini';
@@ -297,7 +298,7 @@ export class EventHelpers {
             mouseEvent.clientY = mouseEvent.pageY = nativeEvent.pageY;
         }
 
-        mouseEvent.button = this.toMouseButton(e.nativeEvent);
+        mouseEvent.button = EventHelpersCommon.toMouseButton(e.nativeEvent);
 
         if (nativeEvent.shiftKey) {
             mouseEvent.shiftKey = nativeEvent.shiftKey;
@@ -333,37 +334,8 @@ export class EventHelpers {
         return dndEvent;
     }
 
-    toMouseButton(nativeEvent: any): number {
-        if (nativeEvent.button !== undefined) {
-            return nativeEvent.button;
-        } else if (nativeEvent.isRightButton || nativeEvent.IsRightButton) {
-            return 2;
-        } else if (nativeEvent.isMiddleButton || nativeEvent.IsMiddleButton) {
-            return 1;
-        }
-
-        return 0;
-    }
-
-    isActuallyMouseEvent(e: Types.TouchEvent | undefined): boolean {
-        if (!e) {
-            return false;
-        }
-
-        const nativeEvent = e as any;
-        if (nativeEvent.button !== undefined) {
-            return true;
-        } else if (nativeEvent.isRightButton || nativeEvent.IsRightButton) {
-            return true;
-        } else if (nativeEvent.isMiddleButton || nativeEvent.IsMiddleButton) {
-            return true;
-        }
-
-        return false;
-    }
-
     isRightMouseButton(e: Types.SyntheticEvent): boolean {
-        return (this.toMouseButton(e.nativeEvent) === 2);
+        return (EventHelpersCommon.toMouseButton(e.nativeEvent) === 2);
     }
 
     // Keyboard events do not inherently hold a position that can be used to show flyouts on keyboard input.
